# Remove padding access from iteration