--- a/include/simdjson/common_defs.h
+++ b/include/simdjson/common_defs.h
@@ -251,6 +251,12 @@ namespace std {
 #endif
 #endif
 
+// Feature flag for partially-implemented "don't require padding" feature
+// TODO remove before 1.0
+#ifndef __SIMDJSON_CHECK_EOF
+# define __SIMDJSON_CHECK_EOF 1
+#endif
+
 #if SIMDJSON_CPLUSPLUS17
 // if we have C++, then fallthrough is a default attribute
 # define simdjson_fallthrough [[fallthrough]]