This removes the "landing pad" On Demand was relying on to detect EOF without bounds checking. Specifically, before we access an index, we check whether there are any further tokens using `try_advance()` or `require_tokens()`.

Remarkably, it does not change the number of instructions executed on Skylake, and does not appear to run slower, either.

This isn't enough to remove padding, but it means the only remaining sources of padding access are the string, number and boolean parsers.