@lemire 2021-03-22
> Remarkably, it does not change the number of instructions executed on Skylake, and does not appear to run slower, either.



How sure are you about this statement?

I am getting an increase in the instruction count on x64...

|   |  instructions before | instructions after  |  % |
|---|---|---|---|
| partial_tweets<simdjson_ondemand>  |  2.41932M  | 2.4388M  | +1% |
| large_random<simdjson_ondemand>  | 598.053M  |  654.053M  | +9% |
|  large_random<simdjson_ondemand_unordered> |  613.053M | 675.053M  | +10% |
| kostya<simdjson_ondemand>  | 631.553M  |  656.718M  | +4% |

AMD Rome (Zen 2), GNU GCC 10.2

I have not pursued the investigation and looking briefly at the numbers, here is what it looks like... you are increasing the instruction count but you are also increasing the number of instructions retired by cycle. In effect, the change you are making appears to be free because we have some slack (that is, the processor can retire more branches per cycle than we retire).

I'd want to look at all the benchmarks and make sure my analysis is correct, and that it generalizes, but I'd like to know what you think.