@jkeiser 2021-03-22
I got that info on skylake. I'll recheck as soon as I get a chance.

I'm ok with free :) I'd really like to circle in on a fix for padding, and this is the first step.