@jkeiser 2021-03-22
@lemire I do have a thought, though. I'm uncomfortable with checking in code that has no immediate benefit. But I want this not to rot and believe we can actually reach a place where padding is not required (possibly even without templating). So how about I do this:

* Add a `#define __SIMDJSON_NO_PADDING_FEATURE_FLAG 1` to simdjson.h
* Put #ifdefs everywhere we check EOF
* Ensure there is no instruction count regression when it's off
* File an issue to remove the flag when we've either got all the other pieces in place, or have decided not to support it.