@lemire 2021-03-22
Here are more complete numbers... It is one interesting case where the number of instructions goes up... with the throughput... so it is not harmful.  There is no apparent increase in the number of mispredicted branches.

|   |  instructions before | instructions after  |  % |
|---|---|---|---|
| partial_tweets<simdjson_ondemand>  |  2.41932M  | 2.4388M  | +1% |
| large_random<simdjson_ondemand>  | 598.053M  |  654.053M  | +9% |
|  large_random<simdjson_ondemand_unordered> |  613.053M | 675.053M  | +10% |
| kostya<simdjson_ondemand>  | 631.553M  |  656.718M  | +4% |
| distinct_user_id<simdjson_ondemand>  | 2.00146M  | 2.02889M  | +1% |
| find_tweet<simdjson_ondemand>  | 1.47803M  |  1.47898M   | +0% |
| top_tweet<simdjson_ondemand>  | 2.08599M  | 2.11351M  | +2% |
| top_tweet<simdjson_ondemand_forward_only>  | 2.29458M  |  2.32196M   | +9% |



|   |  throughput before | throughput after  |  % |
|---|---|---|---|
| partial_tweets<simdjson_ondemand>  |  2.97  GB/s | 3.02 GB/s   | +1% |
| large_random<simdjson_ondemand>  | 0.76 GB/s  | 0.81 GB/s   | +7% |
|  large_random<simdjson_ondemand_unordered> |   0.74 GB/s | 0.78 GB/s  | +5% |
| kostya<simdjson_ondemand>  | 2.28 GB/s   |  2.39 GB/s  | +3% |
| distinct_user_id<simdjson_ondemand>  | 3.50 GB/s   | 3.51 GB/s   | +0% |
| find_tweet<simdjson_ondemand>  | 4.77 GB/s   |  4.81 GB/s  | +1% |
| top_tweet<simdjson_ondemand>  | 3.31 GB/s  |  3.40 GB/s  | +3% |
| top_tweet<simdjson_ondemand_forward_only>  |  3.14 GB/s   | 3.08 GB/s   | -2% |