@lemire 2021-03-22
@jkeiser I think `__SIMDJSON_NO_PADDING_FEATURE_FLAG ` (or `SIMDJSON_NO_PADDING_FEATURE_FLAG`) would be fine as a temporary measure.

A generally good thing is to be able to flip padding on and off and to measure the cost.

I still then we might end up keeping the padded version in the end... if, as I expect, there are benefits.

The performance changes are non-blocking. It may even be that this PR improves performance (somehow) though I'd want to check on different hardware and different compilers.

 But maybe we also want to discuss how we move forward from there. You did not spell out your plan.

I can see several approaches.

1. Make all atom (numbers, true, false...) processing functions 'length' aware.
2. Have two types of atom processing functions. The current ones, and safe ones. At the cost of a branch, you go for unsafe or safe.

These two approaches are probably identical once you go down the compiled binary.

Did you have something else in mind ?

If what you had in mind is what I just described, then this can be implemented nicely.