@jkeiser 2021-03-22
#2 was the approach I was thinking of, yes :)

I want to spend a little more time here, however. Here are the numbers I get for kostya_ondemand:

```
master:
/home/jkeiser/json_benchmark_results/v0.9.1/skylake-clang11.json | grep label
      "label": "[BEST: throughput=  2.25 GB/s doc_throughput=    16 docs/s instructions=   654611050 cycles=   224698482 branch_miss=  461982 cache_miss=10065334 cache_ref=  14033623 items=    524288 avg_time=  60978335 ns]"

PR:
/home/jkeiser/json_benchmark_results/v0.9.1/6/skylake-clang11.json | grep label
      "label": "[BEST: throughput=  2.14 GB/s doc_throughput=    15 docs/s instructions=   712282742 cycles=   236963704 branch_miss=  453647 cache_miss=10327606 cache_ref=  14359086 items=    524288 avg_time=  64264369 ns]"

PR with CHECK_EOF off:
/home/jkeiser/json_benchmark_results/v0.9.1/6/skylake-clang11-nocheckeof.json | grep label
      "label": "[BEST: throughput=  2.15 GB/s doc_throughput=    15 docs/s instructions=   698126953 cycles=   235468189 branch_miss=  467424 cache_miss=10220436 cache_ref=  14040920 items=    524288 avg_time=  63881238 ns]"
```

It could be I just did CHECK_EOF wrong, but the reordering might have caused regressions too. I'll investigate. It'd be nice to have fewer instructions so we have room for other changes.