@lemire 2021-03-23
I suggest merging the new PR https://github.com/simdjson/simdjson/pull/1516 into this PR.