@jkeiser 2021-04-30
A note that I'd really like to explore each change in this patch to see where the slowdown comes from. I have a strong suspicion that the slowdown is related to refactoring / code massaging rather than the design change.