@lemire 2021-05-08
@jkeiser I synced this PR with the main branch (to avoid getting too far off).