@jkeiser 2021-05-20
A thought. Given time pressures it might be best to merge this and do the performance investigation later. There is a very nonzero chance the performance drop is either compiler noise or necessary baggage.