@lemire 2021-05-20
@jkeiser We are getting a new team member of the summer (Nicolas). He is starting new week and I will ask him to run an independent benchmark evaluation covering different compilers and processors.

Otherwise, I am not opposed to merging this into our main branch right now...